feat: added globalStandardize modul and splitted detector into architecture and inference part
