+++ src/AWARE/detection/__init__.py
@@ -1,3 +1,4 @@
-from .multibit_detector import MultibitSTFTMagnitudeDetector, MultibitSTFTMagnitudeDetectorNet
+from .multibit_detector import MultibitSTFTMagnitudeDetector
+from .multibit_detector_net import MultibitSTFTMagnitudeDetectorNetAWARE
 
-__all__ = ['MultibitSTFTMagnitudeDetector', 'MultibitSTFTMagnitudeDetectorNet']
+__all__ = ['MultibitSTFTMagnitudeDetector', 'MultibitSTFTMagnitudeDetectorNetAWARE']
